callbacks: rename package to cref
