+++ internal/cref/cref_test.go
@@ -1,4 +1,4 @@
-package callbacks
+package cref
 
 import (
 	"testing"
